+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added support for reporting regulatory fees
 - Made dividend and brokerage fee account names to use configurable
 - Bumped `apca` dependency to `0.19`